+++ forms-flow-documents/requirements/prod.txt
@@ -16,5 +16,5 @@ markupsafe
 PyJWT
 selenium==3.141.0
 selenium-wire==3.0.0
-nested-lookup==0.2.25
+nested-lookup
 git+https://github.com/AOT-Technologies/forms-flow-ai.git@develop#egg=formsflow_api_utils&subdirectory=forms-flow-api-utils
