+++ forms-flow-documents/src/formsflow_documents/resources/pdf.py
@@ -15,6 +15,7 @@ from formsflow_api_utils.utils import (
 from werkzeug.utils import secure_filename
 
 from formsflow_documents.services import PDFService
+from formsflow_documents.utils import DocUtils
 
 API = Namespace("Form", description="Form")
 
@@ -35,25 +36,23 @@ class FormResourceRenderPdf(Resource):
         use_template = bool(template_name)
 
         template_name = (
-            pdf_service.url_decode(secure_filename(template_name))
+            DocUtils.url_decode(secure_filename(template_name))
             if use_template
             else default_template
         )
 
         template_variable_name = (
-            pdf_service.url_decode(secure_filename(template_variable_name))
+            DocUtils.url_decode(secure_filename(template_variable_name))
             if template_variable_name
             else None
         )
         if not pdf_service.search_template(template_name):
-            raise BusinessException(
-                "Template not found!", HTTPStatus.INTERNAL_SERVER_ERROR
-            )
+            raise BusinessException("Template not found!", HTTPStatus.BAD_REQUEST)
         if template_variable_name and not pdf_service.search_template(
             template_variable_name
         ):
             raise BusinessException(
-                "Template variables not found!", HTTPStatus.INTERNAL_SERVER_ERROR
+                "Template variables not found!", HTTPStatus.BAD_REQUEST
             )
 
         render_data = pdf_service.get_render_data(
@@ -84,57 +83,50 @@ class FormResourceExportPdf(Resource):
 
     @staticmethod
     @auth.require
+    @auth.has_one_of_roles([REVIEWER_GROUP, CLIENT_GROUP])
     @profiletime
     def post(form_id: string, submission_id: string):
         """PDF generation and rendering method."""
         try:
-            if auth.has_one_of_roles([REVIEWER_GROUP, CLIENT_GROUP]):
 
-                timezone = request.args.get("timezone")
-                request_json = request.get_json()
-                template = request_json.get("template")
-                template_variables = request_json.get("templateVars")
-                token = request.headers.get("Authorization")
-                use_template = bool(template)
+            timezone = request.args.get("timezone")
+            request_json = request.get_json()
+            template = request_json.get("template")
+            template_variables = request_json.get("templateVars")
+            token = request.headers.get("Authorization")
+            use_template = bool(template)
 
-                pdf_service = PDFService(form_id=form_id, submission_id=submission_id)
+            pdf_service = PDFService(form_id=form_id, submission_id=submission_id)
 
-                template_name = None
-                template_variable_name = None
+            template_name = None
+            template_variable_name = None
+            if use_template:
+                (
+                    template_name,
+                    template_variable_name,
+                ) = pdf_service.create_template(template, template_variables)
+                current_app.logger.info(template_name)
+                current_app.logger.info(template_variable_name)
+            assert pdf_service.get_render_status(token, template_name) == 200
+            current_app.logger.info("Generating PDF...")
+            result = pdf_service.generate_pdf(
+                timezone, token, template_name, template_variable_name
+            )
+            if result:
                 if use_template:
-                    (
-                        template_name,
-                        template_variable_name,
-                    ) = pdf_service.create_template(template, template_variables)
-                    current_app.logger.info(template_name)
-                    current_app.logger.info(template_variable_name)
-                assert pdf_service.get_render_status(token, template_name) == 200
-                current_app.logger.info("Generating PDF...")
-                result = pdf_service.generate_pdf(
-                    timezone, token, template_name, template_variable_name
-                )
-                if result:
-                    if use_template:
-                        current_app.logger.info("Removing temporary files...")
-                        pdf_service.delete_template(template_name)
-                        if template_variable_name:
-                            pdf_service.delete_template(template_variable_name)
-                    return result
-                response, status = (
-                    {
-                        "message": "Cannot render pdf.",
-                    },
-                    HTTPStatus.BAD_REQUEST,
-                )
-                return response, status
-
+                    current_app.logger.info("Removing temporary files...")
+                    pdf_service.delete_template(template_name)
+                    if template_variable_name:
+                        pdf_service.delete_template(template_variable_name)
+                return result
             response, status = (
                 {
-                    "message": "Permission Denied",
+                    "message": "Cannot render pdf.",
                 },
-                HTTPStatus.FORBIDDEN,
+                HTTPStatus.BAD_REQUEST,
             )
             return response, status
+
         except BusinessException as err:
             current_app.logger.warning(err.error)
             return err.error, err.status_code
