+++ forms-flow-documents/src/formsflow_documents/services/__init__.py
@@ -1,3 +1,3 @@
-"""PDF export helper modules."""
+"""Service layer for Document generation."""
 
 from .pdf import PDFService
